Complete rewrite of the Twig extension
The entire extension is rewritten to be more in-line with the Symfony
Form component. Passing variables to a theme is no longer possible,
but this was properly already broken.

AbstractRendererEngine should be moved to the core package, but
this requires much more thorough testing before this considered
acceptable. The current code is based-of of Symfony's, so it has
been properly tested.